:rocket: Extract context to fix Page deploy error
